Se creo un archivo con constantes para las stacks. Se creo una vista para para el detalle de la movie. Se agrego el parametro navigation desde el componente hijo hasta el padre para poder pasar de vista tocando una tarjeta
