+++ src/components/movie/movie.js
@@ -1,4 +1,4 @@
-import * as React from "react";
+import React from "react";
 import { StyleSheet, Image, View, Text } from "react-native";
 
 export default function Movie({
