+++ src/constants/screens.js
@@ -1,0 +1,4 @@
+export default {
+  search: "Search Movies",
+  movie: "Movie"
+}
