+++ src/screens/movie/ScreenMovie.js
@@ -1,0 +1,13 @@
+import React from "react";
+import Movie from "../../components/movie/movie";
+
+export default function ScreenMovie() {
+  return (
+    <Movie
+      poster={{
+        uri:
+          "https://m.media-amazon.com/images/M/MV5BNDYxNjQyMjAtNTdiOS00NGYwLWFmNTAtNThmYjU5ZGI2YTI1XkEyXkFqcGdeQXVyMTMxODk2OTU@._V1_SX300.jpg",
+      }}
+    />
+  );
+}
